+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 Unreleased
 ----------
 
+- v2: do not override custom API endpoints for zoned requests #806
+
 3.1.46
 ------